+++ src/reweighting.c
@@ -6,7 +6,7 @@
 #include "ccmpred.h"
 #include "reweighting.h"
 
-void calculate_weights(conjugrad_float_t *w, unsigned char *msa, int ncol, int nrow, conjugrad_float_t threshold) {
+void calculate_weights(conjugrad_float_t *w, unsigned char *msa, uint64_t ncol, uint64_t nrow, conjugrad_float_t threshold) {
 	int idthres = (int)ceil(threshold * ncol);
 
 	memset(w, 0, sizeof(conjugrad_float_t) * nrow);
@@ -28,7 +28,7 @@ void calculate_weights(conjugrad_float_t *w, unsigned char *msa, int ncol, int n
 	 	}
 
 		int ids = 0;
-		for(int k = 0; k < ncol; k++) {
+		for(uint64_t k = 0; k < ncol; k++) {
 			if(msa[msa_index(i, k)] == msa[msa_index(j, k)]) {
 				ids++;
 			}
@@ -40,28 +40,28 @@ void calculate_weights(conjugrad_float_t *w, unsigned char *msa, int ncol, int n
 		}
 	}
 
-	for(int i = 0; i < nrow; i++) {
+	for(uint64_t i = 0; i < nrow; i++) {
 		w[i] = 1./(1 + w[i]);
 	}
 
 	conjugrad_float_t wsum = 0;
 	conjugrad_float_t wmin = w[0], wmax = w[0];
-	for(int i = 0; i < nrow; i++) {
+	for(uint64_t i = 0; i < nrow; i++) {
 		conjugrad_float_t wt = w[i];
 		wsum += wt;
 		if(wt > wmax) { wmax = wt; }
 		if(wt < wmin) { wmin = wt; }
 	}
 
-	printf("Reweighted %d sequences with threshold %.1f to Beff=%g weight mean=%g, min=%g, max=%g\n", nrow, threshold, wsum, wsum / nrow, wmin, wmax);
+	printf("Reweighted %ld sequences with threshold %.1f to Beff=%g weight mean=%g, min=%g, max=%g\n", nrow, threshold, wsum, wsum / nrow, wmin, wmax);
 
 }
 
 
-void uniform_weights(conjugrad_float_t *w, int nrow) {
-	for(int i = 0; i < nrow; i++) {
+void uniform_weights(conjugrad_float_t *w, uint64_t nrow) {
+	for(uint64_t i = 0; i < nrow; i++) {
 		w[i] = F1;
 	}
 
-	printf("Using uniform weights on %d sequences\n", nrow);
+	printf("Using uniform weights on %ld sequences\n", nrow);
 }
